+++ packages/gov-ui-kit/CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `RadioGroup` & `Radio` components
+-   Implement `RadioGroup`, `Radio`, and `RadioCard` components
 
 ## [1.0.11] - 2024-02-06
 
@@ -21,7 +21,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Added
 
--   Implement `Link`, `InputNumber` `InputTime`, `TextArea` and `TextAreaRichText` components
+-   Implement `Link`, `InputNumber`, `InputTime`, `TextArea` and `TextAreaRichText` components
 -   Implement Addon element for `InputText` component
 -   Handle size property on `Progress` component
 -   `border-none` Tailwind CSS utility class